[FIX] index_qgram_stree: call to pow fixed on windows
git-svn-id: http://svn.seqan.de/seqan/trunk@13060 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ extras/include/seqan/index/index_qgram_stree.h
@@ -218,7 +218,7 @@ _getNodeByChar(Iter<Index<TText, IndexQGram<TShapeSpec, TIndexSpec> >, VSTree<TS
     childDesc.lastChar = c;
 
     // TODO(esiragusa): Remove call to pow
-    THValue h = pow(ValueSize<TAlphabet>::VALUE, weight(shape) - childDesc.repLen - 1);
+    THValue h = pow(static_cast<double>(ValueSize<TAlphabet>::VALUE), static_cast<int>(weight(shape) - childDesc.repLen - 1));
     childDesc.hash.i1 = childDesc.hash.i1 + ordValue(childDesc.lastChar) * h;
     childDesc.hash.i2 = childDesc.hash.i2 - (ValueSize<TAlphabet>::VALUE - 1 - ordValue(childDesc.lastChar)) * h;
 
@@ -252,7 +252,7 @@ inline bool _getNextNode(Iter<Index<TText, IndexQGram<TShapeSpec, TIndexSpec> >,
     TShape const & shape = indexShape(index);
 
     // TODO(esiragusa): Remove call to pow
-    THValue h = pow(ValueSize<TAlphabet>::VALUE, weight(shape) - value(it).repLen);
+    THValue h = pow(static_cast<double>(ValueSize<TAlphabet>::VALUE), static_cast<int>(weight(shape) - value(it).repLen));
 
     // TODO(esiragusa): Remove workaround for alphabets with quality values
     for (typename ValueSize<TAlphabet>::Type lastChar = ordValue(value(it).lastChar) + 1; lastChar < ValueSize<TAlphabet>::VALUE - 1; ++lastChar)
